Janitoring: address some portability issues

+++ opm/input/eclipse/EclipseState/Grid/EclipseGrid.cpp
@@ -17,9 +17,6 @@
   along with OPM.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include <cstddef>
-#include <stdexcept>
-#include <vector>
 #define _USE_MATH_DEFINES
 
 #include <opm/input/eclipse/EclipseState/Grid/EclipseGrid.hpp>
@@ -61,12 +58,15 @@
 
 #include <algorithm>
 #include <cmath>
+#include <cstddef>
 #include <cstring>
+#include <cstdint>
 #include <functional>
 #include <initializer_list>
 #include <numeric>
 #include <tuple>
-#include <unordered_map>
+#include <stdexcept>
+#include <vector>
 
 #include <fmt/format.h>
 
@@ -1549,7 +1549,7 @@ std::vector<double> EclipseGrid::createDVector(const std::array<int,3>& dims, st
             std::vector<double> volume(this->m_nactive);
 
             #pragma omp parallel for schedule(static)
-            for (std::size_t active_index = 0; active_index < this->m_active_to_global.size(); active_index++) {
+            for (std::int64_t active_index = 0; active_index < static_cast<std::int64_t>(this->m_active_to_global.size()); active_index++) {
                 std::array<double,8> X;
                 std::array<double,8> Y;
                 std::array<double,8> Z;
